pyproject.toml
[project]
name = "trend-engine"
version = "0.1.0"
description = "Multi-source trend aggregator for e-commerce intelligence"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "pytrends>=4.9.2",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
trend-engine = "trend_engine.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
